ALS: Adapt max lux values by observing correction calculation.
Adaptive brightness with an screen sensor which is placed
under screen is heavily dependend on the activated oled
pixels above the sensor.

Therefore we can use unmodified sensor value if the display
shows an black area over the sensor, which acts as as reference
ambient light value.

Now we can use an app which is able to show separate
red/green/blue/white/black values.

Now the automatic brightness is disabled, to get stable results.

By enabling logs the "Corrected reading:" values are observed
and properties are adapted in that way, that with all colors show
the same corrected value.

Change-Id: I9efd703e8400584afad90ea4efec334c53402732